backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alchemy-sdk==0.1.1
alchemy-sdk-py==0.2.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.3
bitarray==3.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.0.1
cloudscraper==1.2.71
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
cytoolz==1.0.1
dataclass-wizard==0.33.0
distlib==0.4.0
dj-database-url==3.1.1
Django==5.1.4
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.9.0
django-filer==3.4.1
django-js-asset==3.1.2
django-pandas==0.6.7
django-phonenumber-field==8.0.0
django-polymorphic==4.5.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
easy-thumbnails==2.10.1
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
filelock==3.20.0
fonttools==4.55.3
frozenlist==1.5.0
gunicorn==25.1.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jsonalias==0.1.1
kiwisolver==1.4.7
lxml==6.0.2
Markdown==3.10
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
phonenumbers==8.13.52
pillow==11.0.0
platformdirs==4.5.0
propcache==0.2.1
psycopg2-binary==2.9.11
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
pyunormalize==16.0.0
pywin32==308
qrcode==7.4.2
regex==2024.11.6
reportlab==4.4.7
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
six==1.17.0
sniffio==1.3.1
solana==0.36.1
solathon==1.0.7
solders==0.23.0
soupsieve==2.6
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.5.1
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.35.4
web3==7.6.1
webencodings==0.5.1
websockets==12.0
whitenoise==6.11.0
yarl==1.18.3
